style(server): undo reorganization of server files
